+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+* Add `is_collection`
+* Cleanup imports in some places
+
 ## 0.7.12 -- 8 Jul 2024
 * Patch `.to_base64()` to `PIL.Image.Image` class to return a base64 encoded string to use in sending images over http requests